Move german-quotation.ts to helpers directory
