+++ client/app/scripts/sir-trevor-blocks/helpers/german-quotation.ts
